Updated OrientationTools to find closest symmetry orientation to reference orientation
Added calc_closest_orientation_configuration_quats to find the closest quaternion configuration (inside or outside the fundamental region) of quats to the ref_quat, this can be used in an effort to "rejoin" orientations separated once mapped to FR. Added an example as well. Should be useful for DSGOD work.